requirements.txt
# Hybrid AI Customer Support System — Dependencies
# Install with:  pip install -r requirements.txt

pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
gradio>=6.12.0
anthropic>=0.20.0        # Optional: only needed if using real LLM
